requirements.txt
requests
streamlit
Pillow
IO